% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_class.R
\name{get_class}
\alias{get_class}
\title{A function to get the classification of each galaxy in a catalog.}
\usage{
get_class(cat, model)
}
\arguments{
\item{cat:}{Data frame with the galaxy catalog. It should have as columns the
distance to the cluster center normalized to the virial radius of each
cluster ('r') and the line-of-sight velocity relative to the cluster
normalized to the line-of-sight velocity of the cluster ('v').}

\item{model:}{Machine Learning model to be used for the classification.}
}
\description{
This function allows you to estimate the orbital classification of each galaxy
 inside a catalog using phase-space information.
}
\examples{
get_class(cat, all_vs_all_KNN)
}
