% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ppsd.R
\name{plot_ppsd}
\alias{plot_ppsd}
\title{A function to plot the projection phase space distribution.}
\usage{
plot_ppsd(
  cl,
  bs,
  ing,
  rin,
  int,
  dens = TRUE,
  cl0 = cl_gal0,
  bs0 = bs_gal0,
  ing0 = in_gal0,
  rin0 = rin_gal0,
  int0 = int_gal0,
  reference = FALSE,
  hist_bins_x = 25,
  hist_bins_y = 20,
  geom = "line"
)
}
\arguments{
\item{cl, }{bs, ing, rin, int: Data frames with the cluster galaxies (cl),
backsplash galaxies (bs), infalling galaxies (ing),
recent infalling galaxies (rin) and interlooper galaxies (int)
of the catalog that will be plotted.
It should have as columns the
distance to the cluster center normalized to the virial radius of each
cluster ('r') and the line-of-sight velocity relative to the cluster
normalized to the line-of-sight velocity of the cluster ('v').}

\item{cl0, }{bs0, ing0, rin0, int0: Data frames with the cluster galaxies (cl),
backsplash galaxies (bs), infalling galaxies (ing),
recent infalling galaxies (rin) and interlooper galaxies (int)
of the reference catalog that will be plotted.
It should have as columns the
distance to the cluster center normalized to the virial radius of each
cluster ('r') and the line-of-sight velocity relative to the cluster
normalized to the line-of-sight velocity of the cluster ('v').}

\item{dens:}{Flag indicating if the density should be plotted. Default = TRUE.}

\item{reference:}{Flag indicating is the properties of the reference catalog should be plotted. 
Default = FALSE.}
}
\description{
This function allows you to plot the projection phase space distribution of a
 dataset.
}
\examples{
plot_ppsd(cl, bs, ing, rin, int)
}
