% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_class.R
\name{get_class}
\alias{get_class}
\title{A function to get the classification of each galaxy in a catalog.}
\usage{
get_class(cat, model, type = "prob", threshold = 0)
}
\arguments{
\item{cat:}{Data frame with the galaxy catalog. It should have as columns the
distance to the cluster center normalized to the virial radius of each
cluster ('r') and the line-of-sight velocity relative to the cluster
normalized to the line-of-sight velocity of the cluster ('v').}

\item{model:}{Machine Learning model to be used for the classification.}

\item{type:}{Output of the machine learning algorithm. Either 'prob' for probabilities or 'raw' for the most probable class. Default 'prob'.}

\item{threshold:}{If type = 'prob' this is the probability threshold to determine the predicted class of a certain galaxy. If threshold = 0, the function returns the full probility vector. Default = 0.}
}
\description{
This function allows you to estimate the orbital classification of each galaxy
 inside a catalog using phase-space information.
}
\examples{
get_class(cat, all_vs_all_KNN)
}
